# internal/assets/openapi.yaml
openapi: 3.0.3
info:
  title: File Management API
  description: API for managing files, folders, tags, and semantic search
  version: 1.0.0
  contact:
    name: RxTech Lab
    url: https://github.com/rxtech-lab/files-management
servers:
  - url: /api
    description: API server
...
paths:
...
  /api/folders:
    get:
      tags:
        - Folders
      summary: List folders
      description: Returns a paginated list of folders with optional filtering
      operationId: listFolders
      parameters:
        - name: keyword
          in: query
          description: Search keyword for folder name
          schema:
            type: string
        - name: parent_id
          in: query
          description: Filter by parent folder ID (omit for root folders)
          schema:
            type: integer
        - name: tag_ids
          in: query
          description: Filter by tag IDs (comma-separated)
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of folders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Folders
      summary: Create folder
      description: Creates a new folder
      operationId: createFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders/tree:
    get:
      tags:
        - Folders
      summary: Get folder tree
      description: Returns the complete folder tree structure
      operationId: getFolderTree
      parameters:
        - name: parent_id
          in: query
          description: Start from this parent folder (omit for full tree from root)
          schema:
            type: integer
      responses:
        '200':
          description: Folder tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderTree'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders/{id}:
    get:
      tags:
        - Folders
      summary: Get folder
      description: Returns a folder by ID with children and tags
      operationId: getFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      responses:
        '200':
          description: Folder details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Folders
      summary: Update folder
      description: Updates an existing folder
      operationId: updateFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFolderRequest'
      responses:
        '200':
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Folders
      summary: Delete folder
      description: Deletes a folder and all its contents recursively
      operationId: deleteFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      responses:
        '204':
          description: Folder deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders/{id}/move:
    post:
      tags:
        - Folders
      summary: Move folder
      description: Moves a folder to a new parent
      operationId: moveFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFolderRequest'
      responses:
        '200':
          description: Folder moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/folders/{id}/tags:
    post:
      tags:
        - Folders
      summary: Add tags to folder
      description: Adds tags to a folder
      operationId: addTagsToFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagIdsRequest'
      responses:
        '200':
          description: Tags added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Folders
      summary: Remove tags from folder
      description: Removes tags from a folder
      operationId: removeTagsFromFolder
      parameters:
        - $ref: '#/components/parameters/FolderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagIdsRequest'
      responses:
        '200':
          description: Tags removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Files
...
  /api/search:
    get:
      tags:
        - Search
      summary: Search files
      description: Searches files using full-text, semantic, or hybrid search
      operationId: searchFiles
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: type
          in: query
          description: Search type
          schema:
            type: string
            enum: [fulltext, semantic, hybrid]
            default: hybrid
        - name: folder_id
          in: query
          description: Limit search to a folder
          schema:
            type: integer
        - name: file_type
          in: query
          description: Filter by file type
          schema:
            $ref: '#/components/schemas/FileType'
        - name: tag_ids
          in: query
          description: Filter by tag IDs (comma-separated)
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Upload
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
...
    FolderId:
      name: id
      in: path
      required: true
      description: Folder ID
      schema:
        type: integer
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message

    # Tags
    Tag:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code (e.g., #FF5733)
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TagIdsRequest:
      type: object
      required:
        - tag_ids
      properties:
        tag_ids:
          type: array
          items:
            type: integer

    # Folders
    Folder:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        description:
          type: string
        parent_id:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        children:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FolderTree:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        parent_id:
          type: integer
          nullable: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/FolderTree'
    CreateFolderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        parent_id:
          type: integer
          nullable: true
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    MoveFolderRequest:
      type: object
      properties:
        parent_id:
          type: integer
          nullable: true
          description: New parent folder ID (null for root)
    FolderListResponse:
      type: object
      required:
        - data
        - total
        - limit
        - offset
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # Files
    FileType:
      type: string
      enum:
        - music
        - photo
        - video
        - document
        - invoice
    ProcessingStatus:
      type: string
      enum:
        - pending
        - processing
        - completed
        - failed
    File:
      type: object
      required:
        - id
        - user_id
        - title
        - s3_key
        - original_filename
        - file_type
        - processing_status
        - has_embedding
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
          description: Parsed text content
        file_type:
          $ref: '#/components/schemas/FileType'
        folder_id:
          type: integer
          nullable: true
        folder:
          $ref: '#/components/schemas/Folder'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        s3_key:
          type: string
        original_filename:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_error:
          type: string
        has_embedding:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SearchResult:
      type: object
      required:
        - file
        - score
      properties:
        file:
          $ref: '#/components/schemas/File'
        score:
          type: number
          format: double
        snippet:
          type: string
    SearchResponse:
      type: object
      required:
        - data
        - total
        - query
        - search_type
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
        query:
          type: string
        search_type:
          type: string

    # Upload
...
security:
  - BearerAuth: []